--- a/admin/src/js/controllers/display-date-time.js
+++ b/admin/src/js/controllers/display-date-time.js
@@ -1,4 +1,4 @@
-const _ = require('underscore');
+const _ = require('lodash');
 const moment = require('moment');
 
 angular.module('controllers').controller('DisplayDateTimeCtrl',
@@ -61,11 +61,11 @@ angular.module('controllers').controller('DisplayDateTimeCtrl',
           reported_date_format: res.reported_date_format,
         };
         $scope.date_formats = standard_date_formats;
-        if (!_.contains($scope.date_formats, res.date_format)) {
+        if (!_.includes($scope.date_formats, res.date_format)) {
           $scope.date_formats.push(res.date_format);
         }
         $scope.datetime_formats = standard_datetime_formats;
-        if (!_.contains($scope.datetime_formats, res.reported_date_format)) {
+        if (!_.includes($scope.datetime_formats, res.reported_date_format)) {
           $scope.datetime_formats.push(res.reported_date_format);
         }