--- a/admin/src/js/controllers/display-languages.js
+++ b/admin/src/js/controllers/display-languages.js
@@ -1,4 +1,5 @@
-const _ = require('lodash');
+const _ = require('lodash/core');
+_.uniq = require('lodash/uniq');
 
 angular.module('controllers').controller('DisplayLanguagesCtrl',
   function (