--- a/admin/src/js/controllers/display-translations.js
+++ b/admin/src/js/controllers/display-translations.js
@@ -1,4 +1,5 @@
-const _ = require('lodash');
+const _ = require('lodash/core');
+_.union = require('lodash/union');
 
 const TRANSLATION_KEYS_OPTION = { doc: {code: 'keys', name: 'Translation Keys'} };
 const DEFAULT_LANGUAGE = 'en';