--- a/admin/src/js/controllers/edit-language.js
+++ b/admin/src/js/controllers/edit-language.js
@@ -1,4 +1,4 @@
-const _ = require('underscore');
+const _ = require('lodash');
 
 angular.module('controllers').controller('EditLanguageCtrl',
   function (