--- a/admin/src/js/controllers/edit-language.js
+++ b/admin/src/js/controllers/edit-language.js
@@ -1,4 +1,4 @@
-const _ = require('lodash');
+const _ = require('lodash/core');
 
 angular.module('controllers').controller('EditLanguageCtrl',
   function (