--- a/admin/src/js/controllers/edit-translation.js
+++ b/admin/src/js/controllers/edit-translation.js
@@ -1,4 +1,4 @@
-const _ = require('lodash');
+const _ = require('lodash/core');
 
 angular.module('controllers').controller('EditTranslationCtrl',
   function (