--- a/admin/src/js/controllers/sms-forms.js
+++ b/admin/src/js/controllers/sms-forms.js
@@ -1,5 +1,5 @@
 const moment = require('moment');
-const _ = require('underscore');
+const _ = require('lodash');
 
 angular.module('controllers').controller('SmsFormsCtrl',
   function(