--- a/admin/src/js/controllers/targets-edit.js
+++ b/admin/src/js/controllers/targets-edit.js
@@ -1,4 +1,4 @@
-const _ = require('lodash');
+const _ = require('lodash/core');
 
 angular.module('controllers').controller('TargetsEditCtrl',
   function (