--- a/admin/src/js/controllers/users.js
+++ b/admin/src/js/controllers/users.js
@@ -1,4 +1,4 @@
-const _ = require('lodash');
+const _ = require('lodash/core');
 
 angular.module('controllers').controller('UsersCtrl',
   function (