--- a/api/src/config.js
+++ b/api/src/config.js
@@ -14,6 +14,10 @@ const translationCache = {};
 let settings = {};
 let transitionsLib;
 
+_.templateSettings = {
+  interpolate: /\{\{(.+?)\}\}/g,
+};
+
 const getMessage = (value, locale) => {
   const _findTranslation = (value, locale) => {
     if (value.translations) {