--- a/api/src/config.js
+++ b/api/src/config.js
@@ -126,7 +126,7 @@ module.exports = {
       (translationCache[locale] && translationCache[locale][key]) ||
       (translationCache.en && translationCache.en[key]) ||
       key;
-    // underscore templates will return ReferenceError if all variables in
+    // lodash templates will return ReferenceError if all variables in
     // template are not defined.
     try {
       return _.template(value)(ctx || {});