--- a/api/src/controllers/bulk-docs.js
+++ b/api/src/controllers/bulk-docs.js
@@ -1,6 +1,6 @@
 const auth = require('../auth');
 const bulkDocs = require('../services/bulk-docs');
-const _ = require('underscore');
+const _ = require('lodash');
 const serverUtils = require('../server-utils');
 
 const requestError = reason => ({
@@ -68,7 +68,7 @@ module.exports = {
 
 // used for testing
 if (process.env.UNIT_TEST_ENV) {
-  _.extend(module.exports, {
+  Object.assign(module.exports, {
     _invalidRequest: invalidRequest,
     _interceptResponse: interceptResponse
   });