--- a/api/src/controllers/people.js
+++ b/api/src/controllers/people.js
@@ -1,4 +1,4 @@
-const _ = require('underscore');
+const _ = require('lodash');
 const db = require('../db');
 const config = require('../config');
 const utils = require('./utils');