--- a/api/src/controllers/places.js
+++ b/api/src/controllers/places.js
@@ -1,4 +1,4 @@
-const _ = require('underscore');
+const _ = require('lodash');
 const config = require('../config');
 const people = require('./people');
 const utils = require('./utils');