@dianabarsan 2020-02-12
Some Search perf improvements (first column represents a Telemetry tag, all measurements were done leveraging Telemetry):

| Desktop (1 million reports)                   | Network | Proposed | process time | Master        | Process time |
|------------------------------------|-----------------|----------|--------------|---------------|--------------|
| `search:contacts:types`            | 30ms            | 35ms     | <0.1s        | 46ms          | <0.1s        |
| `search:reports`                   | 60ms            | 62ms     | <0.1s        | 68ms          | <0.1s        |
| `search:reports:forms`               | 3s              | 3.2s     | 0.2s         | 106s          | 103s         |
| `search:reports:date:forms`          | 27s             | 28s      | 1s           | freeze |              |
| `search:reports:verified`      | 23s             | 26s      | 3s           | freeze |              |
| `search:reports:date:verified`       | 31s             | 33.8s    | 3s           | freeze |              |
| `search:reports:date:forms:verified` | 34s             | 37s      | 3s           | freeze |              |

| Nokia 7 Plus (1 million reports)     | Network  | Proposed | process time | Master | Process time |
|------------------------------------|-----------------|----------|--------------|--------|--------------|
| `search:contacts:types`              | 0.3s            | 0.3s     | <0.1s        | 0.3s   | <0.1s        |
| `search:reports`                     | 0.4s            | 0.3s     | <0.1s        | 0.4s   | <0.1s        |
| `search:reports:forms`               | 19s             | 20s      | 1s           | freeze |              |
| `search:reports:date:forms`          | 121s            | 128s     | 6s           | freeze |              |
| `search:reports:date:forms:verified` | 157s            | 166s     | 9s           | freeze |              |

| Tecno (5k reports)                 | Network | Proposed | process time | Master | Process time |
|------------------------------------|-----------------|----------|--------------|--------|--------------|
| `search:contacts:types`             | 1s              | 1s       | <0.1s        | 1.1s   | <0.1s        |
| `search:reports`                     | 1.8s            | 1.8s     | <0.1s        | 1.8s   | <0.1s        |
| `search:reports:forms`               | 34s             | 34.8     | 0.8s         | 39s    | 5s           |
| `search:reports:date:forms`          | 60s             | 61s      | 1s           | 112s   | 62s          |
| `search:reports:date:forms:verified` | 93s             | 94s      | 1s           | 220s   | 127s         |