@dianabarsan 2020-02-12
@garethbowen , please have a look! 
I just re-read it and it's a bit all over the place, sorry about that. 

These are the replacements I've (hopefully) made: 
| Underscore          | Lodash        |   |   | Lodash/core missing        |
|---------------------|---------------|---|---|----------------------------|
|                     |               |   |   |                            |
| flatten([], true)   | flatten       |   |   | _.includes                 |
| flatten([])         | flattenDeep   |   |   | _.partial (no placeholder) |
| findWhere           | find          |   |   | _.uniq                     |
| _.extend            | Object.assign |   |   | _.uniqBy                   |
| contains            | includes      |   |   | _.reject                   |
| pairs               | toPairs       |   |   | _.chain                    |
| pluck               | map           |   |   | _.template                 |
| pick([], predicate) | pickBy        |   |   | _.templateSettings         |
| uniq([], property)  | uniqBy        |   |   | _.groupBy                  |
| max([], predicate)  | maxBy         |   |   | _.toPairs                  |
| min({[], predicate) | minBy         |   |   | _.minBy                    |
| extendOwn           | assignIn      |   |   | _.union                    |
| where               | filter        |   |   | _.difference               |

Thanks for your time!