@dianabarsan 2020-02-14
Umm, I made a couple of changes in the Search function which maybe should be AT-ed before merging.